refactor: app folders
