+++ src/components/forms/AddTaskForm.tsx
@@ -18,8 +18,8 @@ import { UserWithCheck } from 'types/users';
 import { trimFields } from 'utils/trimFields';
 
 import ControlledFormInput from 'components/ControlledFormInput';
+import CustomPaper from 'components/CustomPaper';
 import ModalWithForm from 'components/ModalWithForm';
-import CustomPaper from 'components/UI/CustomPaper';
 import UserSearchBar from 'components/UserSearchBar';
 
 const AddTaskForm: FC = () => {
