+++ src/components/forms/EditTaskForm.tsx
@@ -17,8 +17,8 @@ import { UserWithCheck } from 'types/users';
 import { trimFields } from 'utils/trimFields';
 
 import ControlledFormInput from 'components/ControlledFormInput';
+import CustomPaper from 'components/CustomPaper';
 import ModalWithForm from 'components/ModalWithForm';
-import CustomPaper from 'components/UI/CustomPaper';
 import UserSearchBar from 'components/UserSearchBar';
 
 const EditTaskForm: FC = () => {
