+++ src/components/layout/footer/RSLogo.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { Box, Link } from '@mui/material';
 import { GRAY_700 } from 'constants/constants';
 
-import RSLogoIcon from '../../assets/icons/logo-rs.svg';
+import RSLogoIcon from '../../../assets/icons/logo-rs.svg';
 
 const RSLogo: FC = () => (
   <Link href="https://rs.school/" color={GRAY_700} target="_blank" rel="noopener noreferrer">
